pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "plotly>=6.1.2",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
    "streamlit>=1.45.1",
    "websocket-client>=1.8.0",
]
